tests/wpt/tests/html/canvas/element/fill-and-stroke-styles/2d.gradient.radial.outside2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.gradient.radial.outside2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.gradient.radial.outside2</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  var g = ctx.createRadialGradient(200, 25, 20, 200, 25, 10);
  g.addColorStop(0, '#0f0');
  g.addColorStop(1, '#f00');
  ctx.fillStyle = g;
  ctx.fillRect(0, 0, 100, 50);

  _assertPixel(canvas, 1,1, 0,255,0,255);
  _assertPixel(canvas, 50,1, 0,255,0,255);
  _assertPixel(canvas, 98,1, 0,255,0,255);
  _assertPixel(canvas, 1,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixelApprox(canvas, 98,25, 0,255,0,255, 1);
  _assertPixel(canvas, 1,48, 0,255,0,255);
  _assertPixelApprox(canvas, 50,48, 0,255,0,255, 1);
  _assertPixel(canvas, 98,48, 0,255,0,255);

});
</script>
